refactor: moved project to default golang layout
