+++ examples/pod-pvc-block-rwop.yaml
