+++ examples/pod-pvc-fs-rwop.yaml
