+++ internal/host/attach.go
