+++ internal/host/delete.go
