+++ internal/host/detach.go
