+++ internal/host/errors.go
